WIP: #5: Start working on tests for FileFsClient, add missing validations

+++ FileFS.Client/FileFsClient.cs
@@ -190,6 +190,11 @@ namespace FileFS.Client
                 throw new InvalidFilenameException(fileName);
             }
 
+            if (!Exists(fileName))
+            {
+                throw new FileNotFoundException(fileName);
+            }
+
             _fileRepository.Delete(fileName);
         }
